Added new OpenGL renderer
Added an OpenGL renderer which may be used in place
of SDL. The renderer interface with the rest of the
code remains largely unchanged. OpenGL appears to
provide much better performance than SDL, especially
when screen scaling is used.

To enable the OpenGL renderer, use the cmake option

cmake -DGRAPHICS_LIBRARY=OpenGL ..

Renamed sdlWindow class to Window.

Renamed sdlColor to ColorRGB.

Added GPU::getBitmapPixel method which is used to
obtain the color of a pixel at position (x,y) in
a tile bitmap.