Fixed GBC boot ROM hanging on logo
The GBC boot ROM is larger than 256 B and was
obscuring the catridge ROM header. The GBC boot
ROM does not use addresses 0x100 to 0x200 where
the cart header is, so the emulator will now ignore
the boot ROM in this range and will instead read
from the cartridge.