Added LR35902 interpreter console
Added a new ConsoleGBC class which is used to draw
a console to the screen as well as handling user
input.

The LR35902 interpreter may be activated by
pressing the ` key at any time. The console may
be used to read/write system memory and cpu
registers as well as executing LR35902 opcodes.