Added controller (gamepad) support
Added support for 360-style gamepads which may now be used in
parallel with keyboard input. Gamepads will be enabled automatically
when connected but the keyboard takes priority if both are in used.

If there is more than one gamepad connected to the system, the
currently active gamepad can be changed by pressing the 'c' key.

If at least one gamepad is connected, pressing F1 will now show
gamepad controls in addition to keyboard controls.