Emulator boots to console if ROM load fails
Emulator now boots to interpreter console if no input ROM is specified or if ROM fails to load correctly.

Console will now refuse to exit back to emulator if no ROM is loaded or if emulator reset fails.

Emulator now continues to refresh the screen even when emulator is paused. This allows text displayed on the screen (e.g. framerate) to continue to update.