Modified breaking after next frame / scanline
Modified the way that breaking after the next frame or scanline functions. These break points now use system clock cycles instead of actual drawing operations. This allows for more precise control as well as the ability to advance to scanlines inside the VBlank interval.

Pressing (or holding) the 'n' key will advance to the next frame (or will render frames until the key is released).