Added filename dropping on exe in Windows
Added support for dropping filename paths on ottergb executable in Windows. If the filename ends with ".cfg" it loads it as a config file (instead of default.cfg), otherwise it attempts to laod the file as a ROM.

Added character map and default palette files to CMake install.

Added error message after failing to load character map.