LCD enable (LCDC bit 7) state change affects PPU
PPU now takes into account if its enable state is being changed
when a program writes to LCDC register bit 7:

If LCD is currently enabled and is being disabled, PPU enters mode 1
and immediately stops drawing. If LCD is disabled and is being
re-enabled, PPU is reset to a new frame and immediatley resumes
drawing. If the state is not being changed, nothing happens.