Updating midi recordings to fix missing notes
Modified the way midi Note On/Off events are added to the output midi file when recording. Only adding notes when an APU channel was enabled caused notes to be missed when the ROM program refreshed the channels without disabling and re-enabling them.

Fixed midi clock multiplier mathematical error which caused output midi files to be unplayable.